+++ packages/react/src/primitives/Select/Select.tsx
@@ -8,6 +8,7 @@ import { IconExpandMore, useIcons } from '../Icon';
 import { ForwardRefPrimitive, Primitive } from '../types';
 import { BaseSelectProps, SelectProps } from '../types/select';
 import { View } from '../View';
+import { useFieldset } from '../Fieldset/useFieldset';
 
 const SelectPrimitive: Primitive<SelectProps, 'select'> = (
   {
@@ -43,6 +44,7 @@ const SelectPrimitive: Primitive<SelectProps, 'select'> = (
     className
   );
   const icons = useIcons('select');
+  const { isFieldsetDisabled } = useFieldset();
 
   return (
     <View className={ComponentClassNames.SelectWrapper}>
@@ -56,7 +58,7 @@ const SelectPrimitive: Primitive<SelectProps, 'select'> = (
             ? DEFAULT_PLACEHOLDER_VALUE
             : defaultValue
         }
-        isDisabled={isDisabled}
+        isDisabled={isFieldsetDisabled ? isFieldsetDisabled : isDisabled}
         multiple={isMultiple}
         size={selectSize}
         required={isRequired}